just moving segy_io file to correct place
